feat(integer-i64): binary-format lint — warn on `f += <integer>` without width
Post-2c, `f += 42` writes 8 bytes to the file (was 4 pre-2c).  For
binary-format files (BigEndian / LittleEndian), that silently
corrupts alignment with older specs expecting i32 fields — the GLB
writer was the flagship regression this migration chased down.

Now the parser warns on `f += <integer_expr>` where the RHS lacks
an explicit width cast (`as i8` / `i16` / `i32` / `u8` / `u16` /
`u32`).  Explicit `as integer` silences the warning as a
self-documenting "yes, I meant 8 bytes" marker.

Implementation is a 10-line diagnostic inside
`src/parser/objects.rs::write_to_file` — added in the else-arm
where we fall through to `self.get_type(val_type)` without a
forced-size alias.  Skips first-pass parsing and `self.default`
(stdlib).

Suppression sites updated:
- `tests/slots.rs::sequential_file_blocks_read_conflict` — two
  intentional 8-byte writes annotated `as integer` (literal
  `0x0102030405060708` and `0x08090a0b0c0d0e0f`).

Tests:
- `binary_write_bare_integer_warns` — asserts the warning fires.
- `binary_write_narrow_cast_silent` — `f += 42 as i32` is silent.
- `binary_write_integer_cast_silent` — `f += 42 as integer` is silent.

Caveat: the parser can't statically know a file's runtime format
(it's set via `f#format = X` at runtime), so the warning fires
even for TextFile writes.  For text-file integer writes, the
explicit `as integer` cast is the idiomatic suppression.

Full suite green.

+++ tests/slots.rs
@@ -107,7 +107,7 @@ fn sequential_file_blocks_read_conflict() {
    assert(f#read(4) as i32 == 0x11223344, \"i32-be-rt\"); }
   delete(\"slots_test_a.bin\");
   {f = file(\"slots_test_a.bin\"); f#format = LittleEndian;
-   f += 0x0102030405060708; assert(f#size == 8, \"long-sz\"); }
+   f += 0x0102030405060708 as integer; assert(f#size == 8, \"long-sz\"); }
   {f = file(\"slots_test_a.bin\"); f#format = LittleEndian;
    assert(f#read(8) as integer == 0x0102030405060708, \"long-rt\"); }
   delete(\"slots_test_a.bin\");
@@ -131,7 +131,7 @@ fn sequential_file_blocks_read_conflict() {
   delete(\"slots_test_b.bin\");
   {f = file(\"slots_test_a.bin\"); f#format = BigEndian;
    f += 0 as u8; f += 1 as u8; f += 0x0203 as u16; f += 0x04050607 as i32;
-   f += 0x08090a0b0c0d0e0f; f += \"Hello world!\"; }
+   f += 0x08090a0b0c0d0e0f as integer; f += \"Hello world!\"; }
   {f = file(\"slots_test_a.bin\"); f#format = LittleEndian;
    assert(f#read(4) as i32 == 0x03020100, \"mixed-4\");
    f#next = 16;
